requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
tqdm>=4.66.0
pyarrow>=15.0.0
streamlit>=1.30.0
